Tidy pkg\system *stat* functions
