+++ pkg/system/stat_solaris.go
@@ -1,12 +1,8 @@
-// +build solaris
-
 package system
 
-import (
-	"syscall"
-)
+import "syscall"
 
-// fromStatT creates a system.StatT type from a syscall.Stat_t type
+// fromStatT converts a syscall.Stat_t type to a system.Stat_t type
 func fromStatT(s *syscall.Stat_t) (*StatT, error) {
 	return &StatT{size: s.Size,
 		mode: uint32(s.Mode),
@@ -14,21 +10,4 @@ func fromStatT(s *syscall.Stat_t) (*StatT, error) {
 		gid:  s.Gid,
 		rdev: uint64(s.Rdev),
 		mtim: s.Mtim}, nil
-}
-
-// FromStatT loads a system.StatT from a syscal.Stat_t.
-func FromStatT(s *syscall.Stat_t) (*StatT, error) {
-	return fromStatT(s)
-}
-
-// Stat takes a path to a file and returns
-// a system.StatT type pertaining to that file.
-//
-// Throws an error if the file does not exist
-func Stat(path string) (*StatT, error) {
-	s := &syscall.Stat_t{}
-	if err := syscall.Stat(path, s); err != nil {
-		return nil, err
-	}
-	return fromStatT(s)
 }
